revised and improved python files
