+++ backend/app/shared/__init__.py
@@ -1,0 +1,2 @@
+"""Shared backend-local copies of DB helper modules."""
+
